R3: finish finyk-domain reuse — inline imports and drop web shims
Replace 1-line re-export shims in apps/web/src/modules/finyk/{domain,lib}/* with direct imports of @sergeant/finyk-domain subpaths. Delete the now-unused shims and the empty domain/ directory. Migrate pure (non-localStorage) tests from apps/web/src/modules/finyk/utils.test.ts into packages/finyk-domain/src/utils.test.ts; keep only lsStats-dependent tests in apps/web. Mark R3 as Done in docs/react-native-migration.md.
