--- apps/web/src/modules/finyk/domain/categories.ts
@@ -1,1 +1,0 @@
-export * from "@sergeant/finyk-domain/domain/categories";
