--- apps/web/src/modules/finyk/domain/selectors.ts
@@ -1,1 +1,0 @@
-export * from "@sergeant/finyk-domain/domain/selectors";
