--- apps/web/src/modules/finyk/domain/transactions.ts
@@ -1,1 +1,0 @@
-export * from "@sergeant/finyk-domain/domain/transactions";
