--- apps/web/src/modules/finyk/domain/types.ts
@@ -1,1 +1,0 @@
-export * from "@sergeant/finyk-domain/domain/types";
