--- apps/web/src/modules/finyk/lib/accounts.ts
@@ -1,1 +1,0 @@
-export * from "@sergeant/finyk-domain/lib/accounts";
