--- apps/web/src/modules/finyk/lib/recurringDetect.ts
@@ -1,1 +1,0 @@
-export * from "@sergeant/finyk-domain/lib/recurringDetect";
