--- apps/web/src/modules/finyk/lib/spending.ts
@@ -1,1 +1,0 @@
-export * from "@sergeant/finyk-domain/lib/spending";
