--- apps/web/src/modules/finyk/lib/transactions.ts
@@ -1,1 +1,0 @@
-export * from "@sergeant/finyk-domain/lib/transactions";
